# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ship Public API",
    "version": "1.0.0",
    "description": "The public, versioned Ship API (PLUGFORGE.MD §2.1/§2.4) — distinct from the\ninternal `/api/*` surface documented at `/api/openapi.json`. Every\nresource here is a typed view over the unified `documents` table.\n\n## Authentication\n\nEvery endpoint except `GET /health` and `GET /openapi.json` requires a\nbearer token: an OAuth access token or a scoped personal access token. See\n`bearerAuth` below.\n\n## Errors\n\nEvery failure response uses one shape (PLUGFORGE.MD §2.5):\n\n```\n{ code, message, details?, request_id }\n```\n\n## Pagination\n\nList endpoints return `{ data, next_cursor }`. `next_cursor` is an opaque,\nkeyset-based cursor — pass it back verbatim as `?cursor=` for the next page;\n`null` means there is no next page."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Public API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A public API bearer token: either an OAuth access token (PF-104, PLUGFORGE.MD §2.2 oauth_tokens) or a scoped personal access token (api_tokens.scopes, PF-107). GET /api/v1/me introspects the authenticated principal (PF-201)."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "unauthorized",
              "forbidden",
              "not_found",
              "validation_failed",
              "rate_limited",
              "server_error"
            ],
            "description": "Machine-readable error code (PLUGFORGE.MD §2.5)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          },
          "details": {
            "type": "object",
            "additionalProperties": {},
            "description": "Additional error context. Shape varies by code — e.g. validation_failed carries fieldErrors, forbidden carries missing_scope."
          },
          "request_id": {
            "type": "string",
            "description": "Correlates this response with the server-side log line for the same request."
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ]
      },
...
      "Issue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Document id."
          },
          "title": {
            "type": "string"
          },
          "document_type": {
            "type": "string",
            "enum": [
              "issue"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "triage",
              "backlog",
              "todo",
              "in_progress",
              "in_review",
              "done",
              "cancelled"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "urgent"
            ]
          },
          "assignee_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 creation timestamp."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 last-updated timestamp."
          }
        },
        "required": [
          "id",
          "title",
          "document_type",
          "state",
          "priority",
          "assignee_id",
          "created_at",
          "updated_at"
        ]
      },
      "IssueList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Issue"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque keyset cursor for the next page (pass back as ?cursor=), or null when this is the last page."
          }
        },
        "required": [
          "data",
          "next_cursor"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/issues": {
      "get": {
        "tags": [
          "Issues"
        ],
        "summary": "List issues",
        "description": "Cursor-paginated list of issue-typed documents in the caller's workspace (PF-201), with state/priority/assignee_id lifted to top-level typed fields. Optionally filtered by ?assignee_id= (PF-205, TRO-414). Requires the issues:read scope.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "assignee_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of issues.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters or cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The bearer token lacks the required scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/issues/{id}": {
      "patch": {
        "tags": [
          "Issues"
        ],
        "summary": "Apply a state transition to an issue",
        "description": "Updates an issue's state field (PF-703, TRO-435) — a deliberately narrow update surface (state only; no title/priority/assignee_id/belongs_to, no \"incomplete children\" confirmation gate — see UpdateIssueRequestSchema's own doc comment). Built for the agent gate's sdk-mode write path (GateShipClient.applyIssueTransition). Requires the issues:write scope.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Document id. A non-UUID value 404s rather than validation-failing."
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "state": {
                    "type": "string",
                    "enum": [
                      "triage",
                      "backlog",
                      "todo",
                      "in_progress",
                      "in_review",
                      "done",
                      "cancelled"
                    ]
                  }
                },
                "required": [
                  "state"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated issue.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Issue"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The bearer token lacks the required scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "No issue with this id exists in the caller's workspace, or the id is malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}